规格参数 判断是否能删除参数

+++ mingrui-shop-service/mingrui-shop-servicexxx/src/main/java/com/baidu/shop/service/impl/SpecificationServiceImpl.java
@@ -1,10 +1,14 @@
 package com.baidu.shop.service.impl;
 
+import com.alibaba.fastjson.JSONObject;
 import com.baidu.shop.base.BaseApiService;
 import com.baidu.shop.base.Result;
 import com.baidu.shop.dto.SpecGroupDTO;
+import com.baidu.shop.dto.SpecParamDTO;
 import com.baidu.shop.entity.SpecGroupEntity;
+import com.baidu.shop.entity.SpecParamEntity;
 import com.baidu.shop.mapper.SpecGroupMapper;
+import com.baidu.shop.mapper.SpecParamMapper;
 import com.baidu.shop.service.SpecificationService;
 import com.baidu.shop.utils.BaiduBeanUtil;
 import com.google.gson.JsonObject;
@@ -18,7 +22,7 @@ import java.util.List;
 /**
  * @ClassName SpecificationServiceImpl
  * @Description: TODO
- * @Author hexiangshen
+ * @Author helianmingbo
  * @Date 2021/1/4
  * @Version V1.0
  **/
@@ -28,6 +32,43 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
     @Resource
     private SpecGroupMapper specGroupMapper;
 
+    @Resource
+    private SpecParamMapper specParamMapper;
+
+
+    @Override//查询规格参数
+    public Result<List<SpecParamEntity>> getSpecParamInfo(SpecParamDTO specParamDTO) {
+        SpecParamEntity specParamEntity = BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class);
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",specParamEntity.getGroupId());
+
+        List<SpecParamEntity> list = specParamMapper.selectByExample(example);
+        return this.setResultSuccess(list);
+    }
+
+    @Transactional
+    @Override
+    public Result<JSONObject> saveSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.insertSelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Transactional
+    @Override
+    public Result<JSONObject> updateSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.updateByPrimaryKey(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Transactional
+    @Override
+    public Result<JSONObject> deleteSpecParam(Integer id) {
+
+        specParamMapper.deleteByPrimaryKey(id);
+        return this.setResultSuccess();
+    }
+
+    //--------------------------------
     @Override
     public Result<List<SpecGroupEntity>> getSpecGroupInfo(SpecGroupDTO specGroupDTO) {
         Example example = new Example(SpecGroupEntity.class);
@@ -43,14 +84,26 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
         return this.setResultSuccess();
     }
 
+    @Transactional
     @Override
     public Result<JsonObject> updateSpecGroup(SpecGroupDTO specGroupDTO) {
         specGroupMapper.updateByPrimaryKeySelective(BaiduBeanUtil.copyProperties(specGroupDTO,SpecGroupEntity.class));
         return this.setResultSuccess();
     }
 
+    @Transactional
     @Override
     public Result<JsonObject> deleteSpecGroup(Integer id) {
+
+        //判断规格足之前是否有规格参数
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",id);
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+        //删除规格组之前需要先判断一下当前规格组下是否有规格参数
+        if(specParamEntities.size() > 0){
+            return this.setResultError("不能删除此参数");
+        }
+
         specGroupMapper.deleteByPrimaryKey(id);
         return this.setResultSuccess();
     }
